[API] Add Queries Generics Application
